Split out Span context
